+++ crates/burn-tensor/src/lib.rs
@@ -50,4 +50,28 @@ mod cube {
             }
         }
     }
+
+}
+
+#[cfg(feature = "cubecl-wgpu")]
+mod cube_wgpu {
+    use crate::backend::{DeviceId, DeviceOps};
+    use cubecl::wgpu::WgpuDevice;
+
+    impl DeviceOps for WgpuDevice {
+        fn id(&self) -> DeviceId {
+            match self {
+                WgpuDevice::DiscreteGpu(index) => DeviceId::new(0, *index as u32),
+                WgpuDevice::IntegratedGpu(index) => DeviceId::new(1, *index as u32),
+                WgpuDevice::VirtualGpu(index) => DeviceId::new(2, *index as u32),
+                WgpuDevice::Cpu => DeviceId::new(3, 0),
+                WgpuDevice::BestAvailable => DeviceId::new(4, 0),
+                // For an existing device, use the 64 bit wgpu device ID as the burn DeviceID.
+                // We're only storing 32 bits, so wrap the the 64 bit value to 32 bits. This
+                // might collide - but a 1 in 4 billion chance seems ok given there's only a few
+                // devices in flight at any time.
+                WgpuDevice::Existing(id) => DeviceId::new(5, (id.inner() % (u32::MAX as u64)) as u32),
+            }
+        }
+    }
 }
