+++ crates/burn-wgpu/Cargo.toml
@@ -23,7 +23,7 @@ cubecl = { workspace = true, features = ["wgpu"] }
 
 burn-jit = { path = "../burn-jit", version = "0.14.0", default-features = false }
 burn-compute = { path = "../burn-compute", version = "0.14.0", default-features = false }
-burn-tensor = { path = "../burn-tensor", version = "0.14.0" }
+burn-tensor = { path = "../burn-tensor", version = "0.14.0", features = ["cubecl-wgpu"] }
 burn-common = { path = "../burn-common", version = "0.14.0" }
 burn-fusion = { path = "../burn-fusion", version = "0.14.0", optional = true }
 burn-cube = { path = "../burn-cube", version = "0.14.0" }
